@@ -17,28 +17,28 @@
 		"@oslojs/encoding": "^1.1.0",
 		"@prisma/client": "^5.21.1",
 		"@sveltejs/adapter-vercel": "^5.4.5",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@tailwindcss/typography": "^0.5.14",
-		"@types/eslint": "^9.6.0",
+		"@sveltejs/kit": "^2.7.2",
+		"@sveltejs/vite-plugin-svelte": "^4",
+		"@tailwindcss/typography": "^0.5.15",
+		"@types/eslint": "^9.6.1",
 		"@types/zxcvbn": "^4.4.5",
 		"arctic": "^2.0.1",
 		"autoprefixer": "^10.4.20",
-		"eslint": "^9.0.0",
+		"eslint": "^9.13.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.36.0",
-		"globals": "^15.0.0",
+		"eslint-plugin-svelte": "^2.46.0",
+		"globals": "^15.11.0",
 		"lucide-svelte": "^0.453.0",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"prettier-plugin-tailwindcss": "^0.6.5",
-		"svelte": "^4.2.7",
-		"svelte-check": "^4.0.0",
+		"prettier": "^3.3.3",
+		"prettier-plugin-svelte": "^3.2.7",
+		"prettier-plugin-tailwindcss": "^0.6.8",
+		"svelte": "^5",
+		"svelte-check": "^4.0.5",
 		"sveltekit-flash-message": "^2.4.4",
-		"tailwindcss": "^3.4.9",
-		"typescript": "^5.0.0",
-		"typescript-eslint": "^8.0.0",
-		"vite": "^5.0.3",
+		"tailwindcss": "^3.4.14",
+		"typescript": "^5.6.3",
+		"typescript-eslint": "^8.11.0",
+		"vite": "^5.4.10",
 		"zxcvbn": "^4.4.2"
 	},
 	"type": "module",
@@ -49,7 +49,7 @@
 		"mode-watcher": "^0.4.1",
 		"prisma": "^5.21.1",
 		"svelte-sonner": "^0.3.28",
-		"sveltekit-superforms": "^2.19.1",
+		"sveltekit-superforms": "^2.20.0",
 		"tailwind-merge": "^2.5.4",
 		"tailwind-variants": "^0.2.1",
 		"zod": "^3.23.8"
