@@ -6,8 +6,11 @@
 	import { Input } from '@/components/ui/input';
 	import { Button } from '@/components/ui/button';
 
-	export let data: SuperValidated<Infer<typeof forgotPasswordSchema>>;
+	interface Props {
+		data: SuperValidated<Infer<typeof forgotPasswordSchema>>;
+	}
 
+	let { data }: Props = $props();
 	const form = superForm(data, {
 		validators: zod(forgotPasswordSchema)
 	});
